[PAYMENT] Update code on Add/Get Card
