+++ src/payment/payment.interface.ts
@@ -1,0 +1,10 @@
+export enum PaymentMethod {
+    Cash = "cash",
+    Card = "card",
+}
+
+export enum PaymentStatus {
+    Pending = "pending",
+    Success = "success",
+    Failed = "failed",
+}
