+++ src/user/util.ts
@@ -4,28 +4,30 @@ import { getDefaultAvatarPaths } from "./avatar";
 import { IStripeCard } from "@/payment/stripe";
 
 type UserProperty = keyof IUser;
+type SafeCard = Partial<Omit<IStripeCard, "paymentMethodId">>;
+type SafeUser = Partial<Omit<IUser, "cards"> & { cards: SafeCard[] }>;
 const safeUserExcludedFields: UserProperty[] = ["password", "role", "updatedAt"];
 
 const safeUserProfileExcludedFields: UserProperty[] = [...safeUserExcludedFields, "addresses"];
 
-function getUserAvatar(user: Partial<IUser>) {
+function getUserAvatar(user: SafeUser) {
     return user.avatar || getDefaultAvatarPaths();
 }
 
 export function getSafeUser(user: IUser) {
-    const safeUser: Partial<IUser> = { ...user };
+    const safeUser: SafeUser = user.toObject();
 
     safeUser.avatar = getUserAvatar(safeUser);
 
     safeUserExcludedFields.forEach((field) => delete safeUser[field]);
 
-    safeUser.cards = getSafeCardInfo(safeUser.cards!);
+    safeUser.cards = getSafeCardInfo(user.cards);
 
     return safeUser;
 }
 
 export function getSafeUserProfile(user: IUser) {
-    const safeUser: Partial<IUser> = { ...user };
+    const safeUser: SafeUser = user.toObject();
 
     safeUserProfileExcludedFields.forEach((field) => delete safeUser[field]);
